Testing new frame spawn
